+++ syscalls.c
@@ -538,11 +538,11 @@ int syscalls_msgRecv(void *ustack)
 {
 	u32 port;
 	msg_t *msg;
-	unsigned int *rid;
+	unsigned long int *rid;
 
 	GETFROMSTACK(ustack, u32, port, 0);
 	GETFROMSTACK(ustack, msg_t *, msg, 1);
-	GETFROMSTACK(ustack, unsigned int *, rid, 2);
+	GETFROMSTACK(ustack, unsigned long int *, rid, 2);
 
 	return proc_recv(port, msg, rid);
 }
@@ -552,11 +552,11 @@ int syscalls_msgRespond(void *ustack)
 {
 	u32 port;
 	msg_t *msg;
-	unsigned int rid;
+	unsigned long int rid;
 
 	GETFROMSTACK(ustack, u32, port, 0);
 	GETFROMSTACK(ustack, msg_t *, msg, 1);
-	GETFROMSTACK(ustack, unsigned int, rid, 2);
+	GETFROMSTACK(ustack, unsigned long int, rid, 2);
 
 	return proc_respond(port, msg, rid);
 }
